Add files via upload
PolyCalcV9 uses a complete redesign of the code structure for greatly improved efficiency and simplicity! Each unit has it's own class created in the Unit_Classes.py folder whilst all functions used to calculate can be found in the Calc_Functions.py file. The output is stored in a simple attackOrderCalc dictionary in which each key corresponds to the attack position and the value is the unit and it's assigned class so all data of interest can be easily accessed.